requirements.txt
streamlit
pandas
openpyxl
supabase
python-dotenv